# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo Application REST-API scheme",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "description": "Просмотр списка пользователей с опциональной пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Список пользователей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Размер страницы с пользователями",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение страницы с пользователями",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение списка пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_users_transport_http.UserDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MulLoMaH_TODO_list_git_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_MulLoMaH_TODO_list_git_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MulLoMaH_TODO_list_git_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "создать нового пользоателя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "CreateUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_MulLoMaH_TODO_list_git_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_MulLoMaH_TODO_list_git_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_MulLoMaH_TODO_list_git_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short human readable message"
                }
            }
        },
...
        "internal_features_users_transport_http.CreateUserRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Ivan Ivanov"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10,
                    "example": "+79998887766"
                }
            }
        },
        "internal_features_users_transport_http.CreateUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "internal_features_users_transport_http.UserDTOResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79998887766"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        }
    }
}